# docs/API_OPENAPI.yaml
openapi: 3.0.3
info:
  title: Auto Investing System Crypto API
  version: 0.1.0-mvp
  description: |
    MVP API specification for the subscription-based AI crypto auto-investing platform.
servers:
  - url: https://api.auto-investing.example.com
    description: Production
  - url: https://staging-api.auto-investing.example.com
    description: Staging
...
security:
  - bearerAuth: []
paths:
...
  /v1/billing/checkout-session:
    post:
      tags: [Billing]
      summary: Create checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan]
              properties:
                plan:
                  $ref: '#/components/schemas/Plan'
      responses:
        '200':
          description: Checkout session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  checkout_url:
                    type: string
                    format: uri
                  session_id:
                    type: string
        '502':
          description: Stripe provider error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/billing/webhooks/stripe:
    post:
      tags: [Billing]
      summary: Process Stripe webhook events and sync subscription status
      security: []
      parameters:
        - in: header
          name: Stripe-Signature
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
                  duplicate:
                    type: boolean
                  status_synced:
                    type: boolean
        '400':
          description: Invalid signature or payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Plan:
      type: string
      enum: [basic, pro, premium]
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true